requirements.txt
requests
speedtest-cli
colorama